% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{approx_env}
\alias{approx_env}
\title{Generates functions for linear interpolation of environmental conditions}
\usage{
approx_env(env_conditions)
}
\description{
Generates functions for linear interpolation of environmental conditions
}
