% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arabian_tractors}
\alias{arabian_tractors}
\title{Number of tractors in the Arab World according to the World Bank}
\format{
A tibble with 40 rows (each corresponding to one year) and 7 columns:
\describe{
\item{year}{Year for the recording}
\item{tractors}{Number of tractors}
}
}
\usage{
arabian_tractors
}
\description{
A dataset showing the increase in tractors in the Arab World. It was retrieved
from https://data.worldbank.org/indicator/AG.AGR.TRAC.NO?end=2009&start=1961&view=chart.
}
\keyword{datasets}
