% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_models.R
\name{Aryani_model}
\alias{Aryani_model}
\title{Secondary Aryani model}
\usage{
Aryani_model(x, xmin, xhalf)
}
\arguments{
\item{x}{Value of the environmental factor.}

\item{xmin}{Minimum value for growth.}

\item{xhalf}{Value where gamma = 0.5}
}
\value{
The corresponding gamma factor.
}
\description{
Secondary cardinal parameter model as defined by Aryani et al. (2015).
}
