% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_isothermal_growth.R
\name{bilinear_lag}
\alias{bilinear_lag}
\title{Bilinear model with lag}
\usage{
bilinear_lag(times, logN0, mu, lambda)
}
\arguments{
\item{times}{Numeric vector of storage times}

\item{logN0}{Initial log microbial count}

\item{mu}{Maximum specific growth rate (in ln CFU/\link{t})}

\item{lambda}{Lag phase duration}
}
\description{
Bilinear model with lag
}
