% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_isothermal_growth.R
\name{bilinear_stationary}
\alias{bilinear_stationary}
\title{Bilinear model with stationary phase}
\usage{
bilinear_stationary(times, logN0, mu, logNmax)
}
\arguments{
\item{times}{Numeric vector of storage times}

\item{logN0}{Initial log microbial count}

\item{mu}{Maximum specific growth rate (in ln CFU/\link{t})}

\item{logNmax}{Maximum log microbial count}
}
\description{
Bilinear model with stationary phase
}
