% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biogrowth-package.R
\docType{package}
\name{biogrowth-package}
\alias{biogrowth}
\alias{biogrowth-package}
\title{biogrowth: Modelling of Population Growth}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Modelling of population growth under static and dynamic environmental conditions. Includes functions for model fitting and making prediction under isothermal and dynamic conditions. The methods (algorithms & models) are based on predictive microbiology (See Perez-Rodriguez and Valero (2012, ISBN:978-1-4614-5519-6)).
}
\author{
\strong{Maintainer}: Alberto Garre \email{garre.alberto@gmail.com} (\href{https://orcid.org/0000-0002-4404-3550}{ORCID})

Authors:
\itemize{
  \item Jeroen Koomen \email{jeroen.koomen@wur.nl}
  \item Heidy den Besten \email{heidy.denbesten@wur.nl}
  \item Marcel Zwietering \email{marcel.zwietering@wur.nl}
}

}
\keyword{internal}
