% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_models.R
\name{calculate_gammas}
\alias{calculate_gammas}
\title{Calculates every gamma factor}
\usage{
calculate_gammas(this_t, env_func, sec_models)
}
\arguments{
\item{this_t}{Storage time}

\item{env_func}{A list of functions (generated using \code{approxfun}) that
give the value of each environmental function for some storage time.}

\item{sec_models}{A nested list describing the secondary models.}
}
\value{
A vector of gamma factors (one per environmental factor).
}
\description{
A helper function for \code{\link{predict_dynamic_growth}} that
calculates the value of every gamma factor corresponding to some
storage time.
}
