% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_secondary_models.R
\name{calculate_gammas_secondary}
\alias{calculate_gammas_secondary}
\title{Gamma factors for fitting secondary models}
\usage{
calculate_gammas_secondary(sec_model_names, my_data, secondary_models)
}
\description{
Gamma factors for fitting secondary models
}
