% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meatinfo_primary.R
\name{check_primary_pars}
\alias{check_primary_pars}
\title{Basic check of parameters for primary models}
\usage{
check_primary_pars(model_name, pars)
}
\arguments{
\item{model_name}{Model identifier}

\item{pars}{A named list of model parameters}
}
\value{
If there is no error, the model function.
}
\description{
Checks that: the model name is correct, the right number of model
parameters have been defined and that the parameters have the right names
}
