% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_stochastic_growth.R
\name{check_stochastic_pars}
\alias{check_stochastic_pars}
\title{Model definition checks for predict_stochastic_growth}
\usage{
check_stochastic_pars(model_name, pars, corr_matrix)
}
\arguments{
\item{model_name}{Character describing the primary growth model.}

\item{pars}{A tibble describing the parameter uncertainty (see details).}

\item{corr_matrix}{Correlation matrix of the model parameters. Defined in the
same order as in \code{pars}. A diagonal matrix by default
(uncorrelated parameters).}
}
\description{
Does several checks of the model parameters. Besides those by 
check_primary_pars, it checks that corr_matrix is square, that pars and
corr_matrix have compatible dimensions, and that pars has the correct names.
}
