% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{coef.FitDynamicGrowth}
\alias{coef.FitDynamicGrowth}
\title{Fitted parameters of a dynamic fit}
\usage{
\method{coef}{FitDynamicGrowth}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitDynamicGrowth}.}

\item{...}{ignored}
}
\description{
Fitted parameters of a dynamic fit
}
