% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{coef.FitDynamicGrowthMCMC}
\alias{coef.FitDynamicGrowthMCMC}
\title{Fitted parameters of an MCMC fit}
\usage{
\method{coef}{FitDynamicGrowthMCMC}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitDynamicGrowthMCMC}.}

\item{...}{ignored}
}
\description{
Fitted parameters of an MCMC fit
}
