% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{coef.FitIsoGrowth}
\alias{coef.FitIsoGrowth}
\title{Fitted parameters of an isothermal fit}
\usage{
\method{coef}{FitIsoGrowth}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitIsoGrowth}.}

\item{...}{ignored}
}
\description{
Fitted parameters of an isothermal fit
}
