% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{coef.FitMultipleDynamicGrowth}
\alias{coef.FitMultipleDynamicGrowth}
\title{Fitted parameters of a global fit}
\usage{
\method{coef}{FitMultipleDynamicGrowth}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitMultipleDynamicGrowth}.}

\item{...}{ignored}
}
\description{
Fitted parameters of a global fit
}
