% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{coef.FitMultipleGrowthMCMC}
\alias{coef.FitMultipleGrowthMCMC}
\title{Fitted parameters of a global fit using MCMC}
\usage{
\method{coef}{FitMultipleGrowthMCMC}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitMultipleGrowthMCMC}.}

\item{...}{ignored}
}
\description{
Fitted parameters of a global fit using MCMC
}
