% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{coef.FitSecondaryGrowth}
\alias{coef.FitSecondaryGrowth}
\title{Fitted parameters of a secondary model fit}
\usage{
\method{coef}{FitSecondaryGrowth}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitSecondaryGrowth}.}

\item{...}{ignored}
}
\description{
Fitted parameters of a secondary model fit
}
