% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareIsoGrowth_class.R
\name{ComparisonIsoGrowth}
\alias{ComparisonIsoGrowth}
\alias{plot.ComparisonIsoGrowth}
\alias{coef.ComparisonIsoGrowth}
\alias{print.ComparisonIsoGrowth}
\alias{summary.ComparisonIsoGrowth}
\title{ComparisonIsoGrowth class}
\usage{
\method{plot}{ComparisonIsoGrowth}(x, y, ..., type = 1, add_trend = TRUE)

\method{coef}{ComparisonIsoGrowth}(object, ...)

\method{print}{ComparisonIsoGrowth}(x, ..)

\method{summary}{ComparisonIsoGrowth}(object, ...)
}
\arguments{
\item{x}{an instance of ComparisonIsoGrowth}

\item{y}{ignored}

\item{...}{ignored}

\item{type}{if type==1, the plot compares the model predictions. If type ==2,
the plot compares the parameter estimates. If type==3, the plot shows the residuals}

\item{object}{an instance of ComparisonIsoGrowth}
}
\description{
The \code{ComparisonIsoGrowth} class contains AA
}
\section{Functions}{
\itemize{
\item \code{plot.ComparisonIsoGrowth}: illustrations comparing the fitted models

\item \code{coef.ComparisonIsoGrowth}: table of parameter estimates

\item \code{print.ComparisonIsoGrowth}: print of the model comparison

\item \code{summary.ComparisonIsoGrowth}: summary table for the comparison
}}

