% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conditions_pH_temperature}
\alias{conditions_pH_temperature}
\title{Conditions during a dynamic growth experiment}
\format{
A tibble with 4 rows and 3 columns:
\describe{
    \item{time}{elapsed time}
    \item{temperature}{temperature}
    \item{pH}{pH}
}
}
\usage{
conditions_pH_temperature
}
\description{
A dataset to demonstrate the use of fit_dynamic_growth. The observations 
environmental conditions are described in conditions_pH_temperature.
}
\keyword{datasets}
