% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_coupled_growth.R
\name{cost_coupled_onestep}
\alias{cost_coupled_onestep}
\title{Residuals of the coupled Baranyi model}
\usage{
cost_coupled_onestep(p, this_data, known)
}
\arguments{
\item{p}{a numeric vector of model parameters. Must have entries \code{logN0}, \code{logNmax}, \code{logC0}, \code{b} and \code{Tmin}}

\item{this_data}{a tibble (or data.frame) with three columns: \code{logN} (microbial concentration; in logCFU/TIME),
\code{temp} the temperature and \code{time} the storage time}

\item{known}{a numeric vector of known model parameters}
}
\value{
the vector of model residuals
}
\description{
Residuals of the coupled Baranyi model
}
