% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_coupled_growth.R
\name{cost_coupled_twosteps}
\alias{cost_coupled_twosteps}
\title{Cost for the coupled model fitted in two-steps}
\usage{
cost_coupled_twosteps(p, this_data, weight = NULL, known)
}
\arguments{
\item{p}{numeric vector (or list) of model parameters. Must have entries \code{logC0}, \code{b} and \code{Tmin}}

\item{weight}{type of weights to apply. Either \code{NULL} (no weights; default),
\code{sd} (standard deviation) or \code{mean} (mean value).}

\item{my_d}{tibble (or data.frame) of data. It must have one column named \code{temp} (temperature),
one named \code{lambda} (specific growth rate; in ln CFU/TIME) and one named \code{mu}
(specific growth rate; in ln CFU/TIME).}
}
\value{
vector of weighted residuals
}
\description{
Cost for the coupled model fitted in two-steps
}
