% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{deviance.FitDynamicGrowth}
\alias{deviance.FitDynamicGrowth}
\title{Deviance of a dynamic fit}
\usage{
\method{deviance}{FitDynamicGrowth}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitDynamicGrowth}.}

\item{...}{ignored}
}
\description{
Deviance of a dynamic fit
}
