% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{deviance.FitDynamicGrowthMCMC}
\alias{deviance.FitDynamicGrowthMCMC}
\title{Deviance of an MCMC fit}
\usage{
\method{deviance}{FitDynamicGrowthMCMC}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitDynamicGrowthMCMC}.}

\item{...}{ignored}
}
\description{
Returns the deviance of the best fit, calculated as the sum of
squared residuals.
}
