% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{deviance.FitIsoGrowth}
\alias{deviance.FitIsoGrowth}
\title{Deviance of an isothermal fit}
\usage{
\method{deviance}{FitIsoGrowth}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitIsoGrowth}.}

\item{...}{ignored}
}
\description{
Deviance of an isothermal fit
}
