% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{deviance.FitMultipleDynamicGrowth}
\alias{deviance.FitMultipleDynamicGrowth}
\title{Deviance of a global fit}
\usage{
\method{deviance}{FitMultipleDynamicGrowth}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitMultipleDynamicGrowth}.}

\item{...}{ignored}
}
\description{
Deviance of a global fit
}
