% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{deviance.FitMultipleGrowthMCMC}
\alias{deviance.FitMultipleGrowthMCMC}
\title{Deviance of a global fit using MCMC}
\usage{
\method{deviance}{FitMultipleGrowthMCMC}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitMultipleGrowthMCMC}.}

\item{...}{ignored}
}
\description{
Returns the deviance of the best fit, calculated as the sum of
squared residuals.
}
