% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{deviance.FitSecondaryGrowth}
\alias{deviance.FitSecondaryGrowth}
\title{Deviance of a cardinal fit}
\usage{
\method{deviance}{FitSecondaryGrowth}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitSecondaryGrowth}.}

\item{...}{ignored}
}
\description{
Deviance of a cardinal fit
}
