% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_cardinal}
\alias{example_cardinal}
\title{Growth rates obtained for several growth experiments}
\format{
A data frame with 64 rows and 3 variables:
\describe{
\item{temperature}{storage temperature (ºC)}
\item{pH}{pH of the media}
\item{mu}{specific growth rate (log10 CFU/h)}
}
}
\usage{
example_cardinal
}
\description{
An example dataset illustrating the requirements of the
\code{\link[=fit_secondary_growth]{fit_secondary_growth()}} function.
}
\keyword{datasets}
