% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_coupled_onestep}
\alias{example_coupled_onestep}
\title{Example data for two-steps fitting of the Baranyi-Ratkowsky model}
\format{
A tibble with three columns:
\itemize{
\item temp: the treatment temperature
\item time: the elapsed time of the sample
\item logN: the (decimal) log microbial concentration
}
}
\usage{
example_coupled_onestep
}
\description{
This dataset serve as an example of the data input for \code{fit_coupled_growth} using the
one-step mode.
}
\keyword{datasets}
