% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_coupled_twosteps}
\alias{example_coupled_twosteps}
\title{Example data for two-steps fitting of the Baranyi-Ratkowsky model}
\format{
A tibble with three columns:
\itemize{
\item temp: the treatment temperature
\item mu: the value of mu estimated at each temperature
\item lambda: the value of lambda estimated at each temperature
}
}
\usage{
example_coupled_twosteps
}
\description{
This dataset serve as an example of the data input for \code{fit_coupled_growth} using the
two-steps mode.
}
\keyword{datasets}
