% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dynamic_growth}
\alias{example_dynamic_growth}
\title{Microbial growth under dynamic conditions}
\format{
A data frame with 30 rows and 2 variables:
\describe{
  \item{time}{elapsed time}
  \item{logN}{log microbial count}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
example_dynamic_growth
}
\description{
An example dataset illustrating the requirements of the
\code{\link{fit_dynamic_growth}} function.
}
\keyword{datasets}
