% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_env_conditions}
\alias{example_env_conditions}
\title{Environmental conditions during a dynamic experiment}
\format{
A data frame with 3 rows and 3 variables:
\describe{
  \item{time}{elapsed time}
  \item{temperature}{storage temperature}
  \item{aw}{water activity}
}
}
\usage{
example_env_conditions
}
\description{
An example dataset illustrating the requirements of the
\code{\link{fit_dynamic_growth}} function.
}
\keyword{datasets}
