% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_od}
\alias{example_od}
\title{Example data for TTD calculation and the serial-dilution method}
\format{
A tibble with 97 rows and 61 columns:
\itemize{
\item the first column (\code{time}) presents the time of the reading
\item the remaining columns present the OD recorded on each well. They are codified as
condition + _ + number of dilutions, according to \code{\link[=get_TTDs]{get_TTDs()}}.
}
}
\usage{
example_od
}
\description{
This dataset serve as an example of the data input for \code{\link[=get_TTDs]{get_TTDs()}}.
}
\keyword{datasets}
