% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_secondary_pars}
\alias{extract_secondary_pars}
\title{A helper to build the secondary models}
\usage{
extract_secondary_pars(this_p, known_pars, sec_model_names)
}
\description{
A helper to build the secondary models
}
