% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_coupled_growth.R
\name{fit_coupled_growth}
\alias{fit_coupled_growth}
\title{Growth fitting considering link between mu and lambda}
\usage{
fit_coupled_growth(
  fit_data,
  start,
  known,
  mode = "two_steps",
  weight = "sd",
  ...,
  logbase_mu = exp(1),
  logbase_logN = 10
)
}
\arguments{
\item{fit_data}{a tibble (or data.frame) with the data for the fit. The content must
be different depending on the fitting mode (see Details).}

\item{start}{a numeric vector of initial guesses for the parameter estimates}

\item{known}{a numeric vector of known mode parameters}

\item{mode}{the type of model fitting approach. Either \code{two_steps} (fitted from the
values of \code{mu} and \code{lambda}) or \code{one_step} (fitted from logN)}

\item{weight}{weights to apply for the \code{two_steps} fit. Either \code{NULL} (no weights),
\code{sd} (standard deviation; default) or \code{mean} (mean value).}

\item{...}{ignored}

\item{logbase_mu}{AA}

\item{logbase_logN}{AA}
}
\description{
Growth fitting considering link between mu and lambda
}
