% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dynamic_growth.R
\name{fit_dynamic_growth}
\alias{fit_dynamic_growth}
\title{Fit dynamic growth models}
\usage{
fit_dynamic_growth(
  fit_data,
  env_conditions,
  starting_point,
  known_pars,
  sec_model_names,
  ...
)
}
\description{
Fit dynamic growth models
}
