% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dynamic_growth.R
\name{fit_MCMC_growth}
\alias{fit_MCMC_growth}
\title{Fit growth models using MCMC}
\usage{
fit_MCMC_growth(
  fit_data,
  env_conditions,
  starting_point,
  known_pars,
  sec_model_names,
  niter,
  ...
)
}
\description{
Fit growth models using MCMC
}
