% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twofold_dilution.R
\name{fit_serial_dilution}
\alias{fit_serial_dilution}
\title{Serial-fold dilution method}
\usage{
fit_serial_dilution(
  TTD_data,
  start,
  dil_factor = 2,
  mode = "intercept",
  logN_det = NULL,
  logN_dil0 = NULL
)
}
\arguments{
\item{TTD_data}{a tibble (or data.frame) with the TTD observed for different dilutions.
It must have two columns: \code{TTD} (the TTD) and \code{dil} the number of serial dilutions.}

\item{dil_factor}{dilution factor. By default, 2}

\item{mode}{one of "intercept" (serial dilution method with a generic intercept; default)
or "lambda" (able to estimate also the value of the lag phase duration)}

\item{logN_det}{log10 microbial concentration at the detection OD (only for mode = "lambda")}

\item{logN_dil0}{log10 microbial concentration at wells where dilution = 0 (only for mode = "lambda")}
}
\description{
Model fitting by the serial-fold dilution method
}
