% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{fitted.FitDynamicGrowth}
\alias{fitted.FitDynamicGrowth}
\title{Fitted values of dynamic fit}
\usage{
\method{fitted}{FitDynamicGrowth}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitDynamicGrowth}.}

\item{...}{ignored}
}
\description{
Fitted values of dynamic fit
}
