% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{fitted.FitDynamicGrowthMCMC}
\alias{fitted.FitDynamicGrowthMCMC}
\title{Fitted values of MCMC fit}
\usage{
\method{fitted}{FitDynamicGrowthMCMC}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitDynamicGrowthMCMC}.}

\item{...}{ignored}
}
\description{
Fitted values of MCMC fit
}
