% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{fitted.FitIsoGrowth}
\alias{fitted.FitIsoGrowth}
\title{Fitted values of static fit}
\usage{
\method{fitted}{FitIsoGrowth}(object, ...)
}
\arguments{
\item{object}{an instance of \code{itIsoGrowth }.}

\item{...}{ignored}
}
\description{
Fitted values of static fit
}
