% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{fitted.FitMultipleDynamicGrowth}
\alias{fitted.FitMultipleDynamicGrowth}
\title{Fitted values of global fit}
\usage{
\method{fitted}{FitMultipleDynamicGrowth}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitMultipleDynamicGrowth}.}

\item{...}{ignored}
}
\value{
A tibble with 3 columns: time (storage time), exp (experiment 
identifier) and fitted (fitted value).
}
\description{
Fitted values of global fit
}
