% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{fitted.FitMultipleGrowthMCMC}
\alias{fitted.FitMultipleGrowthMCMC}
\title{Fitted values of global fit with MCMC}
\usage{
\method{fitted}{FitMultipleGrowthMCMC}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitMultipleGrowthMCMC}.}

\item{...}{ignored}
}
\value{
A tibble with 3 columns: time (storage time), exp (experiment 
identifier) and fitted (fitted value).
}
\description{
Fitted values of global fit with MCMC
}
