% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{fitted.FitSecondaryGrowth}
\alias{fitted.FitSecondaryGrowth}
\title{Fitted values of cardinal fit}
\usage{
\method{fitted}{FitSecondaryGrowth}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitSecondaryGrowth}.}

\item{...}{ignored}
}
\description{
The fitted values are returned in the same scale as the one
used for the fitting (sqrt, log or none).
}
