% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_models.R
\name{full_Ratkowski}
\alias{full_Ratkowski}
\title{Full Ratkowsky model}
\usage{
full_Ratkowski(x, xmin, xmax, c)
}
\arguments{
\item{x}{Value of the environmental factor.}

\item{xmin}{Minimum value for growth}

\item{xmax}{Maximum value for growth}

\item{c}{Parameter defining the speed of the decline}
}
\description{
Gamma model adapted from the one by Ratkowsky et al. (1983).
}
