% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalGrowthComparison_class.R
\name{get_all_predictions}
\alias{get_all_predictions}
\title{A helper for making the plots}
\usage{
get_all_predictions(model)
}
\arguments{
\item{model}{An instance of FitMultipleDynamicGrowth}
}
\description{
A helper for making the plots
}
