% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dynamic_growth.R
\name{get_dyna_residuals}
\alias{get_dyna_residuals}
\title{Residuals of dynamic prediction}
\usage{
get_dyna_residuals(
  this_p,
  fit_data,
  env_conditions,
  known_pars,
  sec_model_names
)
}
\description{
Residuals of dynamic prediction
}
