% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_iso_growth.R
\name{get_iso_residuals}
\alias{get_iso_residuals}
\title{Residuals of isothermal prediction}
\usage{
get_iso_residuals(this_p, fit_data, model_name, known_pars)
}
\description{
Residuals of isothermal prediction
}
