% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_secondary_models.R
\name{get_secondary_residuals}
\alias{get_secondary_residuals}
\title{Residuals of secondary models}
\usage{
get_secondary_residuals(
  this_p,
  my_data,
  known_pars,
  sec_model_names,
  transformation
)
}
\description{
Residuals of secondary models
}
