% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalGrowthComparison_class.R
\name{GlobalGrowthComparison}
\alias{GlobalGrowthComparison}
\alias{coef.GlobalGrowthComparison}
\alias{summary.GlobalGrowthComparison}
\alias{print.GlobalGrowthComparison}
\alias{plot.GlobalGrowthComparison}
\title{GlobalGrowthComparison class}
\usage{
\method{coef}{GlobalGrowthComparison}(object, ...)

\method{summary}{GlobalGrowthComparison}(object, ...)

\method{print}{GlobalGrowthComparison}(x, ..)

\method{plot}{GlobalGrowthComparison}(x, y, ..., type = 1, add_trend = TRUE)
}
\arguments{
\item{object}{an instance of GlobalGrowthComparison}

\item{...}{ignored}

\item{x}{an instance of GlobalGrowthComparison}

\item{y}{ignored}

\item{type}{if type==1, the plot compares the model predictions. If type ==2,
the plot compares the parameter estimates. If type==3, the plot shows the residuals}
}
\description{
The \code{GlobalGrowthComparison} class contains AA
}
\section{Functions}{
\itemize{
\item \code{coef.GlobalGrowthComparison}: table of parameter estimates

\item \code{summary.GlobalGrowthComparison}: summary table for the comparison

\item \code{print.GlobalGrowthComparison}: print of the model comparison

\item \code{plot.GlobalGrowthComparison}: illustrations comparing the fitted models
}}

