% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{greek_tractors}
\alias{greek_tractors}
\title{Number of tractors in Greece according to the World Bank}
\format{
A tibble with 46 rows (each corresponding to one year) and 7 columns:
\describe{
\item{year}{Year for the recording}
\item{tractors}{Number of tractors}
}
}
\usage{
greek_tractors
}
\description{
A dataset showing the increase in tractors in Greece. It was retrieved
from https://data.worldbank.org/indicator/AG.AGR.TRAC.NO?end=2009&start=1961&view=chart.
}
\keyword{datasets}
