% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{growth_pH_temperature}
\alias{growth_pH_temperature}
\title{Example of dynamic growth}
\format{
A tibble with 20 rows and 2 columns:
\describe{
    \item{time}{elapsed time}
    \item{logN}{decimal logarithm of the population size}
}
}
\usage{
growth_pH_temperature
}
\description{
A dataset to demonstrate the use of fit_dynamic_growth. The values of the 
environmental conditions are described in conditions_pH_temperature.
}
\keyword{datasets}
