% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{growth_salmonella}
\alias{growth_salmonella}
\title{Growth of Salmonella spp in broth}
\format{
A tibble with 21 rows and 2 columns:
\describe{
\item{time}{elapsed time in hours.}
\item{logN}{observed population size (log CFU/g).}
}
}
\usage{
growth_salmonella
}
\description{
An example dataset to illustrate \code{\link[=fit_isothermal_growth]{fit_isothermal_growth()}}. It
describes the growth of Salmonella spp. in broth. It was retrieved from
ComBase (ID: B092_10).
}
\keyword{datasets}
