% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareIsoGrowth_class.R
\name{GrowthComparison}
\alias{GrowthComparison}
\alias{plot.GrowthComparison}
\alias{coef.GrowthComparison}
\alias{print.GrowthComparison}
\alias{summary.GrowthComparison}
\title{GrowthComparison class}
\usage{
\method{plot}{GrowthComparison}(x, y, ..., type = 1, add_trend = TRUE)

\method{coef}{GrowthComparison}(object, ...)

\method{print}{GrowthComparison}(x, ..)

\method{summary}{GrowthComparison}(object, ...)
}
\arguments{
\item{x}{an instance of GrowthComparison}

\item{y}{ignored}

\item{...}{ignored}

\item{type}{if type==1, the plot compares the model predictions. If type ==2,
the plot compares the parameter estimates. If type==3, the plot shows the residuals}

\item{object}{an instance of GrowthComparison}
}
\description{
The \code{GrowthComparison} class contains AA
}
\section{Functions}{
\itemize{
\item \code{plot.GrowthComparison}: illustrations comparing the fitted models

\item \code{coef.GrowthComparison}: table of parameter estimates

\item \code{print.GrowthComparison}: print of the model comparison

\item \code{summary.GrowthComparison}: summary table for the comparison
}}

