% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_models.R
\name{inhibitory_model}
\alias{inhibitory_model}
\title{Secondary model for inhibitory compounds}
\usage{
inhibitory_model(x, MIC, alpha)
}
\arguments{
\item{x}{Value of the environmental factor (in principle, concentration of compound).}

\item{MIC}{Minimum Inhibitory Concentration}

\item{alpha}{shape factor of the miodel}
}
\value{
The corresponding gamma factor.
}
\description{
Secondary model for the effect of inhibitory compounds.
}
