% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{is.DynamicGrowth}
\alias{is.DynamicGrowth}
\title{Test of DynamicGrowth object}
\usage{
is.DynamicGrowth(x)
}
\arguments{
\item{x}{object to be checked.}
}
\value{
A boolean specifying whether \code{x} is of class
        \code{DynamicGrowth}
}
\description{
Tests if an object is of class \code{DynamicGrowth}.
}
