% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{is.FitDynamicGrowth}
\alias{is.FitDynamicGrowth}
\title{Test of FitDynamicGrowth object}
\usage{
is.FitDynamicGrowth(x)
}
\arguments{
\item{x}{object to be checked.}
}
\value{
A boolean specifying whether \code{x} is of class
        \code{FitDynamicGrowth}
}
\description{
Tests if an object is of class \code{FitDynamicGrowth}.
}
