% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{is.FitDynamicGrowthMCMC}
\alias{is.FitDynamicGrowthMCMC}
\title{Test of FitDynamicGrowthMCMC object}
\usage{
is.FitDynamicGrowthMCMC(x)
}
\arguments{
\item{x}{object to be checked.}
}
\value{
A boolean specifying whether \code{x} is of class
        \code{FitDynamicGrowthMCMC}
}
\description{
Tests if an object is of class \code{FitDynamicGrowthMCMC}.
}
