% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{is.FitIsoGrowth}
\alias{is.FitIsoGrowth}
\title{Test of FitIsoGrowth object}
\usage{
is.FitIsoGrowth(x)
}
\arguments{
\item{x}{object to be checked.}
}
\value{
A boolean specifying whether \code{x} is of class
\code{FitIsoGrowth}
}
\description{
Tests if an object is of class \code{FitIsoGrowth}.
}
