% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{is.FitSecondaryGrowth}
\alias{is.FitSecondaryGrowth}
\title{Test of FitSecondaryGrowth object}
\usage{
is.FitSecondaryGrowth(x)
}
\arguments{
\item{x}{object to be checked.}
}
\value{
A boolean specifying whether \code{x} is of class
        \code{FitSecondaryGrowth}
}
\description{
Tests if an object is of class \code{FitSecondaryGrowth}.
}
