% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{is.GlobalGrowthFit}
\alias{is.GlobalGrowthFit}
\title{Test of GlobalGrowthFit object}
\usage{
is.GlobalGrowthFit(x)
}
\arguments{
\item{x}{object to be checked.}
}
\value{
A boolean specifying whether \code{x} is of class GlobalGrowthFit
}
\description{
Tests if an object is of class \link{GlobalGrowthFit}
}
