% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{is.GrowthFit}
\alias{is.GrowthFit}
\title{Test of GrowthFit object}
\usage{
is.GrowthFit(x)
}
\arguments{
\item{x}{object to be checked.}
}
\value{
A boolean specifying whether \code{x} is of class GrowthFit
}
\description{
Tests if an object is of class \link{GrowthFit}
}
