% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{is.GrowthPrediction}
\alias{is.GrowthPrediction}
\title{Test of GrowthPrediction object}
\usage{
is.GrowthPrediction(x)
}
\arguments{
\item{x}{object to be checked.}
}
\value{
A boolean specifying whether \code{x} is of class GrowthPrediction
}
\description{
Tests if an object is of class \link{GrowthPrediction}
}
