% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{is.GrowthUncertainty}
\alias{is.GrowthUncertainty}
\title{Test of GrowthUncertainty object}
\usage{
is.GrowthUncertainty(x)
}
\arguments{
\item{x}{object to be checked.}
}
\value{
A boolean specifying whether \code{x} is of class GrowthUncertainty
}
\description{
Tests if an object is of class \link{GrowthUncertainty}
}
