% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{is.IsothermalGrowth}
\alias{is.IsothermalGrowth}
\title{Test of IsothermalGrowth object}
\usage{
is.IsothermalGrowth(x)
}
\arguments{
\item{x}{object to be checked.}
}
\value{
A boolean specifying whether \code{x} is of class
\code{IsothermalGrowth}
}
\description{
Tests if an object is of class \code{IsothermalGrowth}.
}
