% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{is.StochasticGrowth}
\alias{is.StochasticGrowth}
\title{Test of StochasticGrowth object}
\usage{
is.StochasticGrowth(x)
}
\arguments{
\item{x}{object to be checked.}
}
\value{
A boolean specifying whether \code{x} is of class
        \code{StochasticGrowth}
}
\description{
Tests if an object is of class \code{StochasticGrowth}.
}
