% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_isothermal_growth.R
\name{iso_Baranyi}
\alias{iso_Baranyi}
\title{Isothermal Baranyi model}
\usage{
iso_Baranyi(times, logN0, mu, lambda, logNmax)
}
\arguments{
\item{times}{Numeric vector of storage times}

\item{logN0}{Initial log microbial count}

\item{mu}{Maximum specific growth rate}

\item{lambda}{Lag phase duration}

\item{logNmax}{Maximum log microbial count}
}
\value{
Numeric vector with the predicted microbial count.
}
\description{
Baranyi growth model as defined by Baranyi and Roberts (1994).
}
