% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsothermalGrowth_class.R
\name{IsothermalGrowth}
\alias{IsothermalGrowth}
\alias{print.IsothermalGrowth}
\alias{plot.IsothermalGrowth}
\title{IsothermalGrowth class}
\usage{
\method{print}{IsothermalGrowth}(x, ...)

\method{plot}{IsothermalGrowth}(x, y = NULL, ..., line_col = "black", line_size = 1, line_type = "solid")
}
\arguments{
\item{x}{The object of class \code{IsothermalGrowth} to plot.}

\item{...}{ignored.}

\item{y}{ignored}

\item{line_col}{Aesthetic parameter to change the colour of the line,
see: \code{\link[=geom_line]{geom_line()}}}

\item{line_size}{Aesthetic parameter to change the thickness of the line,
see: \code{\link[=geom_line]{geom_line()}}}

\item{line_type}{Aesthetic parameter to change the type of the line,
takes numbers (1-6) or strings ("solid") see: \code{\link[=geom_line]{geom_line()}}}
}
\description{
The \code{IsothermalGrowth} class contains the results of a growth prediction
under isothermal conditions. Its constructor is \code{\link[=predict_isothermal_growth]{predict_isothermal_growth()}}.

It is a subclass of list with the items:
\itemize{
\item simulation: A tibble with the model simulation.
\item model: The name of the model used for the predictions.
\item pars: A list with the values of the model parameters.
}
}
\section{Functions}{
\itemize{
\item \code{print.IsothermalGrowth}: print of the model

\item \code{plot.IsothermalGrowth}: plot of the predicted growth curve.
}}

