% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lambda_to_Q0}
\alias{lambda_to_Q0}
\title{Q0 from lag phase duration}
\usage{
lambda_to_Q0(lambda, mu)
}
\arguments{
\item{lambda}{Duration of the lag phase.}

\item{mu}{Specific growth rate in the exponential phase (log10(units)/[time]).}
}
\description{
Convenience function to calculate the value of Q0 for the Baranyi model from
the duration of the lag phase
}
