% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_isothermal_growth.R
\name{logistic_model}
\alias{logistic_model}
\title{Logistic growth model}
\usage{
logistic_model(times, logN0, mu, lambda, C)
}
\arguments{
\item{times}{Numeric vector of storage times}

\item{logN0}{Initial log microbial count}

\item{mu}{Maximum specific growth rate (in ln CFU/[t])}

\item{lambda}{Lag phase duration}

\item{C}{Difference between \code{logN0} and the maximum log-count.}
}
\value{
Numeric vector with the predicted microbial count
}
\description{
Logistic growth model
}
