% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_isothermal_growth.R
\name{loglinear_model}
\alias{loglinear_model}
\title{Loglinear model}
\usage{
loglinear_model(times, logN0, mu)
}
\arguments{
\item{times}{Numeric vector of storage times}

\item{logN0}{Initial log microbial count}

\item{mu}{Maximum specific growth rate (in ln CFU/\link{t})}
}
\description{
Loglinear model
}
