% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_guesses.R
\name{make_guess_factor}
\alias{make_guess_factor}
\title{Initial guesses for the secondary model of one factor}
\usage{
make_guess_factor(fit_data, sec_model, factor, logbase = c("log10", "natural"))
}
\arguments{
\item{fit_data}{Tibble with the data used for the fit. It must have
one column with the observed growth rate (named \code{mu} by default; can be
changed using the "formula" argument) and as many columns
as needed with the environmental factors.}

\item{sec_model}{character defining the secondary model equation according
to \code{\link[=secondary_model_data]{secondary_model_data()}}}

\item{factor}{character defining the environmental factor}
}
\description{
Initial guesses for the secondary model of one factor
}
