% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiple_conditions}
\alias{multiple_conditions}
\title{Environmental conditions during several dynamic experiments}
\format{
A nested list with two elements, each one corresponding to one experiment.
Each element is a data.frame with three columns:
\itemize{
\item time: elapsed time
\item temperature: observed temperature
\item pH: observed pH
}
}
\usage{
multiple_conditions
}
\description{
This dataset is paired with \link{multiple_counts} to illustrate the global
fitting of \code{\link[=fit_growth]{fit_growth()}}.
}
\keyword{datasets}
