% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiple_counts}
\alias{multiple_counts}
\title{Population growth observed in several dynamic experiments}
\format{
A nested list with two elements, each one corresponding to one experiment.
Each element is a data.frame with two columns:
\itemize{
\item time: elapsed time
\item logN: log10 of the microbial concentration
}
}
\usage{
multiple_counts
}
\description{
This dataset is paired with \link{multiple_conditions} to illustrate the global
fitting of \code{\link[=fit_growth]{fit_growth()}}.
}
\keyword{datasets}
