% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_plotting.R
\name{plot.FitDynamicGrowthMCMC}
\alias{plot.FitDynamicGrowthMCMC}
\title{Plot of FitDynamicGrowthMCMC Object}
\usage{
\method{plot}{FitDynamicGrowthMCMC}(
  x,
  y = NULL,
  ...,
  add_factor = NULL,
  ylims = NULL,
  label_y1 = "logN",
  label_y2 = add_factor,
  line_col = "black",
  line_size = 0.5,
  line_type = 1,
  point_col = "black",
  point_size = 0.5,
  point_shape = 16
)
}
\arguments{
\item{x}{The object of class \code{FitDynamicGrowthMCMC} to plot.}

\item{y}{ignored}

\item{...}{additional arguments passed to \code{plot}.}

\item{add_factor}{whether to plot also one environmental factor.
If \code{NULL} (default), no environmenta factor is plotted. If set
to one character string that matches one entry of x$env_conditions,
that condition is plotted in the secondary axis}

\item{ylims}{A two dimensional vector with the limits of the primary y-axis.}

\item{label_y1}{Label of the primary y-axis.}

\item{label_y2}{Label of the secondary y-axis.}

\item{line_col}{Aesthetic parameter to change the colour of the line geom in the plot, see: \code{\link{geom_line}}}

\item{line_size}{Aesthetic parameter to change the thickness of the line geom in the plot, see: \code{\link{geom_line}}}

\item{line_type}{Aesthetic parameter to change the type of the line geom in the plot, takes numbers (1-6) or strings ("solid") see: \code{\link{geom_line}}}

\item{point_col}{Aesthetic parameter to change the colour of the point geom, see: \code{\link{geom_point}}}

\item{point_size}{Aesthetic parameter to change the size of the point geom, see: \code{\link{geom_point}}}

\item{point_shape}{Aesthetic parameter to change the shape of the point geom, see: \code{\link{geom_point}}}
}
\value{
An instance of \code{ggplot}.
}
\description{
Plots the predicted log microbial count for an
instance of \code{FitDynamicGrowthMCMC}.
}
