% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_plotting.R
\name{plot.FitIsoGrowth}
\alias{plot.FitIsoGrowth}
\title{Plot of FitIsoGrowth Object}
\usage{
\method{plot}{FitIsoGrowth}(x, y = NULL, ...)
}
\arguments{
\item{x}{The object of class \code{FitIsoGrowth} to plot.}

\item{y}{ignored}

\item{...}{additional arguments passed to \code{plot}.}
}
\value{
An instance of \code{ggplot}.
}
\description{
Plots the predicted log microbial count for an
instance of \code{FitIsoGrowth}.
}
