% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_plotting.R
\name{plot.FitSecondaryGrowth}
\alias{plot.FitSecondaryGrowth}
\title{Plot of FitSecondaryGrowth}
\usage{
\method{plot}{FitSecondaryGrowth}(x, y = NULL, ..., which = 1, add_trend = FALSE)
}
\arguments{
\item{x}{An instance of FitSecondaryGrowth.}

\item{y}{ignored.}

\item{...}{ignored.}

\item{which}{A numeric with the type of plot. 1 for obs versus predicted (default),
2 for gamma curve}

\item{add_trend}{Whether to add a trend line (only for which=2)}
}
\description{
Plot of FitSecondaryGrowth
}
