% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_plotting.R
\name{plot.TimeDistribution}
\alias{plot.TimeDistribution}
\title{Plot of TimeDistribution Object}
\usage{
\method{plot}{TimeDistribution}(x, y = NULL, ..., bin_width = 0.5)
}
\arguments{
\item{x}{The object of class \code{TimeDistribution} to plot.}

\item{y}{ignored}

\item{...}{additional arguments passed to \code{plot}.}

\item{bin_width}{A number that specifies the width of a bin in the histogram, see: \code{\link{geom_histogram}}}
}
\value{
An instance of \code{ggplot}.
}
\description{
Plots the predicted log microbial count for an
instance of \code{TimeDistribution}.
}
