% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_coupled_growth.R
\name{pred_coupled_baranyi}
\alias{pred_coupled_baranyi}
\title{Predictions of the coupled Baranyi model}
\usage{
pred_coupled_baranyi(p, temp, times)
}
\arguments{
\item{p}{a numeric vector of model parameters. Must have entries \code{logN0}, \code{logNmax}, \code{logC0}, \code{b} and \code{Tmin}}

\item{temp}{a numeric vector of temperature values}

\item{times}{a numeric vector of time points for the prediction}
}
\value{
a numeric vector of predicted logN (in log CFU/TIME)
}
\description{
Predictions of the coupled Baranyi model
}
