% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_coupled_growth.R
\name{pred_lambda}
\alias{pred_lambda}
\title{Prediction of lambda for the coupled model}
\usage{
pred_lambda(p, temp)
}
\arguments{
\item{p}{numeric vector (or list) of model parameters. Must have entries \code{logC0}, \code{b} and \code{Tmin}}

\item{temp}{numeric vector of temperatures}
}
\value{
the values of lambda
}
\description{
Prediction of lambda for the coupled model
}
