% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_coupled_growth.R
\name{pred_sqmu}
\alias{pred_sqmu}
\title{Prediction of the square root of mu for the coupled model}
\usage{
pred_sqmu(p, temp)
}
\arguments{
\item{p}{numeric vector (or list) of model parameters. Must have entries \code{b} and \code{Tmin}}

\item{temp}{numeric vector of temperatures}
}
\value{
the values of the square root of mu (in ln CFU/TIME)
}
\description{
Prediction of the square root of mu for the coupled model
}
