% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{predict.FitDynamicGrowthMCMC}
\alias{predict.FitDynamicGrowthMCMC}
\title{Model prediction for dynamic fit using MCMC}
\usage{
\method{predict}{FitDynamicGrowthMCMC}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an instance of FitDynamicGrowthMCMC}

\item{newdata}{a tibble describing the environmental conditions (as \code{env_conditions})
in \code{\link{predict_dynamic_growth}}. 
If \code{NULL} (default), uses the same conditions as those for fitting.}

\item{...}{ignored}
}
\description{
Model prediction for dynamic fit using MCMC
}
