% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{predict.FitIsoGrowth}
\alias{predict.FitIsoGrowth}
\title{Model prediction for static fit}
\usage{
\method{predict}{FitIsoGrowth}(object, times = NULL, ...)
}
\arguments{
\item{object}{an instance of FitIsoGrowth}

\item{times}{numeric vector describing the time points for the prediction.
If \code{NULL} (default), uses the same points as those used for fitting.}

\item{...}{ignored}
}
\description{
Model prediction for static fit
}
