% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{predict.FitMultipleDynamicGrowth}
\alias{predict.FitMultipleDynamicGrowth}
\title{Model predictions of FitMultipleDynamicGrowth}
\usage{
\method{predict}{FitMultipleDynamicGrowth}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{Instance of FitMultipleDynamicGrowth}

\item{newdata}{a tibble describing the environmental conditions (as \code{env_conditions})
in \code{\link{fit_multiple_growth}}. 
If \code{NULL} (default), uses the same conditions as those for fitting.}

\item{...}{ignored}
}
\value{
A table with 3 columns: time (storage time), logN (observed count),
and exp (name of the experiment).
}
\description{
Model predictions of FitMultipleDynamicGrowth
}
