% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{predict.FitSecondaryGrowth}
\alias{predict.FitSecondaryGrowth}
\title{Model predictions for FitSecondaryGrowth}
\usage{
\method{predict}{FitSecondaryGrowth}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an instance of FitSecondaryGrowth}

\item{newdata}{}

\item{...}{ignored}
}
\description{
Model predictions for FitSecondaryGrowth
}
