% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dynamic_growth.R
\name{predict_dynamic_growth}
\alias{predict_dynamic_growth}
\title{Growth under dynamic conditions}
\usage{
predict_dynamic_growth(times, env_conditions, primary_pars, secondary_models)
}
\description{
Growth under dynamic conditions
}
