% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_predict.R
\name{predict_growth}
\alias{predict_growth}
\title{Prediction of microbial growth}
\usage{
predict_growth(
  times,
  primary_model,
  environment = c("constant", "dynamic"),
  secondary_model = NULL,
  env_conditions = NULL,
  type = c("deterministic", "interval"),
  n_sims = NULL,
  par_corr = NULL,
  ...,
  check = TRUE,
  logbase = c("natural", "10"),
  formula = . ~ time
)
}
\description{
AA
}
