% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_isothermal_growth.R
\name{predict_isothermal_growth}
\alias{predict_isothermal_growth}
\title{Isothermal microbial growth}
\usage{
predict_isothermal_growth(model_name, times, model_pars)
}
\description{
Isothermal microbial growth
}
