% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_MCMC_growth.R
\name{predict_MCMC_growth}
\alias{predict_MCMC_growth}
\title{Stochastic growth of MCMC fit}
\usage{
predict_MCMC_growth(MCMCfit, times, env_conditions, niter)
}
\description{
Stochastic growth of MCMC fit
}
