% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMCMC.R
\name{predictMCMC}
\alias{predictMCMC}
\title{Generic for calculating predictions with uncertainty from fits}
\usage{
predictMCMC(model, times, env_conditions, niter, newpars = NULL, ...)
}
\arguments{
\item{model}{Fit object}

\item{times}{see specific methods for each class}

\item{env_conditions}{see specific methods for each class}

\item{niter}{see specific methods for each class}

\item{newpars}{see specific methods for each class}
}
\description{
Generic for calculating predictions with uncertainty from fits
}
