% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMCMC_coupled.R
\name{predictMCMC_coupled}
\alias{predictMCMC_coupled}
\title{Generic for calculating predictions with uncertainty from fits}
\usage{
predictMCMC_coupled(model, niter, newdata = NULL)
}
\arguments{
\item{model}{An instance of FitCoupledGrowth}

\item{niter}{number of MC simulations}

\item{newdata}{a tibble (or data.frame) with two columns (time and temperature)
for the prediction. By default, \code{NULL} (the fitting conditions)}
}
\description{
Generic for calculating predictions with uncertainty from fits
}
