% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meatinfo_primary.R
\name{primary_model_data}
\alias{primary_model_data}
\title{Metainformation of primary growth models}
\usage{
primary_model_data(model_name = NULL)
}
\arguments{
\item{model_name}{The name of the model or \code{NULL} (default).}
}
\value{
If model_name is \code{NULL}, returns a character string with the available models.
If is a valid identifier, it returns a list with metainformation about the model.
If model_name name is not a valid identifier, raises an error.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Provides different types of meta-data about the primary growth models included
in biogrowth. This information is the basis of the automatic checks, and can also
help in the definition of models for \code{\link[=predict_growth]{predict_growth()}} and \code{\link[=fit_growth]{fit_growth()}}.
}
