% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refrigeratorSpain}
\alias{refrigeratorSpain}
\title{Temperature recorded in refrigerators}
\format{
A tibble with three columns:
\itemize{
\item time: elapsed time in hours
\item A1: temperature observed in refrigerator "1"
\item A2: temperature observed in refrigerator "2"
}
}
\usage{
refrigeratorSpain
}
\description{
This dataset includes the temperature recorded in refrigerators in households
of the Catalonia region. The data was published as part of Jofre et al. (2019)
Domestic refrigerator temperatures in Spain: Assessment of its impact on the
safety and shelf-life of cooked meat products. Food Research International,
126, 108578. And was kindly provided by the original authors of the study.
}
\keyword{datasets}
