% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{residuals.FitDynamicGrowth}
\alias{residuals.FitDynamicGrowth}
\title{Residuals of FitDynamicGrowth}
\usage{
\method{residuals}{FitDynamicGrowth}(object, ...)
}
\arguments{
\item{object}{Instance of FitDynamicGrowth}

\item{...}{ignored}
}
\description{
Residuals of FitDynamicGrowth
}
