% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{residuals.FitDynamicGrowthMCMC}
\alias{residuals.FitDynamicGrowthMCMC}
\title{Residuals of FitDynamicGrowthMCMC}
\usage{
\method{residuals}{FitDynamicGrowthMCMC}(object, ...)
}
\arguments{
\item{object}{Instance of FitDynamicGrowthMCMC.}

\item{...}{ignored.}
}
\description{
Residuals of FitDynamicGrowthMCMC
}
