% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{residuals.FitIsoGrowth}
\alias{residuals.FitIsoGrowth}
\title{Residuals of FitIsoGrowth object}
\usage{
\method{residuals}{FitIsoGrowth}(object, ...)
}
\arguments{
\item{object}{Instance of FitIsoGrowth}

\item{...}{ignored}
}
\description{
Residuals of FitIsoGrowth object
}
