% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{residuals.FitMultipleDynamicGrowth}
\alias{residuals.FitMultipleDynamicGrowth}
\title{Residuals of FitMultipleDynamicGrowth}
\usage{
\method{residuals}{FitMultipleDynamicGrowth}(object, ...)
}
\arguments{
\item{object}{Instance of FitMultipleDynamicGrowth}

\item{...}{ignored}
}
\description{
Residuals of FitMultipleDynamicGrowth
}
