% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{residuals.FitMultipleGrowthMCMC}
\alias{residuals.FitMultipleGrowthMCMC}
\title{Residuals of FitMultipleGrowthMCMC}
\usage{
\method{residuals}{FitMultipleGrowthMCMC}(object, ...)
}
\arguments{
\item{object}{Instance of FitMultipleGrowthMCMC}

\item{...}{ignored}
}
\value{
A table with 4 columns: time (storage time), logN (observed count),
exp (name of the experiment) and res (residual).
}
\description{
Residuals of FitMultipleGrowthMCMC
}
