% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{residuals.FitSecondaryGrowth}
\alias{residuals.FitSecondaryGrowth}
\title{Residuals of a FitSecondaryGrowth object}
\usage{
\method{residuals}{FitSecondaryGrowth}(object, ...)
}
\arguments{
\item{object}{Instance of FitSecondaryGrowth}

\item{...}{ignored}
}
\description{
Residuals of a FitSecondaryGrowth object
}
