% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_coupled_growth.R
\name{residuals_lambda}
\alias{residuals_lambda}
\title{Residuals for lambda for the coupled model}
\usage{
residuals_lambda(p, my_d)
}
\arguments{
\item{p}{numeric vector (or list) of model parameters. Must have entries \code{logC0}, \code{b} and \code{Tmin}}

\item{my_d}{tibble (or data.frame) of data. It must have one column named \code{temp} (temperature)
and one named \code{lambda} (specific growth rate; in ln CFU/TIME).}
}
\value{
vector of residuals
}
\description{
Residuals for lambda for the coupled model
}
