% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_models.R
\name{Rossoaw_model}
\alias{Rossoaw_model}
\title{Secondary Rosso model for water activity}
\usage{
Rossoaw_model(x, xmin)
}
\arguments{
\item{x}{Value of the environmental factor (in principle, aw).}

\item{xmin}{Minimum value for growth (in principle, aw).}
}
\value{
The corresponding gamma factor.
}
\description{
Secondary model for water activity as defined by Aryani et al. (2001).
}
