% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metainfo_secondary.R
\name{secondary_model_data}
\alias{secondary_model_data}
\title{Metainformation of secondary growth models}
\usage{
secondary_model_data(model_name = NULL)
}
\arguments{
\item{model_name}{The name of the model or \code{NULL} (default).}
}
\value{
If model_name is \code{NULL}, returns a character string with the available models.
If is a valid identifier, it returns a list with metainformation about the model.
If model_name name is not a valid identifier, raises an error.
}
\description{
Metainformation of secondary growth models
}
