% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_guesses.R
\name{show_guess_primary}
\alias{show_guess_primary}
\title{Plot of the initial guess for growth under constant environmental conditions}
\usage{
show_guess_primary(
  fit_data,
  model,
  guess,
  logbase = c("natural", "10"),
  formula = logN ~ time
)
}
\arguments{
\item{fit_data}{Tibble (or data.frame) of data for the fit. It must have two columns, one with
the elapsed time (\code{time} by default) and another one with the decimal logarithm
of the populatoin size (\code{logN} by default). Different column names can be
defined using the \code{formula} argument.}

\item{guess}{Named vector with the initial guess of the model parameters}

\item{formula}{an object of class "formula" describing the x and y variables.
\code{logN ~ time} as a default.}

\item{model_name}{Character defining the primary growth model as per \code{\link[=primary_model_data]{primary_model_data()}}}
}
\value{
A \code{\link[=ggplot]{ggplot()}} comparing the model prediction against the data
}
\description{
Compares the prediction corresponding to a guess of the parameters of the primary
model against experimental data
}
