% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StochasticGrowth_class.R
\name{StochasticGrowth}
\alias{StochasticGrowth}
\alias{print.StochasticGrowth}
\alias{plot.StochasticGrowth}
\title{StochasticGrowth class}
\usage{
\method{print}{StochasticGrowth}(x, ...)

\method{plot}{StochasticGrowth}(
  x,
  y = NULL,
  ...,
  line_col = "black",
  line_size = 0.5,
  line_type = "solid",
  ribbon80_fill = "grey",
  ribbon90_fill = "grey",
  alpha80 = 0.5,
  alpha90 = 0.4
)
}
\arguments{
\item{x}{The object of class \code{StochasticGrowth} to plot.}

\item{...}{ignored.}

\item{y}{ignored}

\item{line_col}{Aesthetic parameter to change the colour of the line geom in the plot, see: \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}}

\item{line_size}{Aesthetic parameter to change the thickness of the line geom in the plot, see: \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}}

\item{line_type}{Aesthetic parameter to change the type of the line geom in the plot, takes numbers (1-6) or strings ("solid") see: \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}}

\item{ribbon80_fill}{fill colour for the space between the 10th and 90th quantile, see: \code{\link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon()}}}

\item{ribbon90_fill}{fill colour for the space between the 5th and 95th quantile, see: \code{\link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon()}}}

\item{alpha80}{transparency of the ribbon aesthetic for the space between the 10th and 90th quantile. Takes a value between 0 (fully transparant) and 1 (fully opaque)}

\item{alpha90}{transparency of the ribbon aesthetic for the space between the 5th and 95th quantile. Takes a value between 0 (fully transparant) and 1 (fully opaque).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

The class \link{StochasticGrowth} has been deprecated by
class \link{GrowthUncertainty}, which provides less misleading name.

Still, it is still returned if the deprecated \code{\link[=predict_stochastic_growth]{predict_stochastic_growth()}} is called.

The \code{StochasticGrowth} class contains the results of a growth prediction
under isothermal conditions considering parameter unceratinty. Its constructor
is \code{\link[=predict_stochastic_growth]{predict_stochastic_growth()}}.

It is a subclass of list with the items:
\itemize{
\item sample: parameter sample used for the calculations.
\item simulations: growth curves predicted for each parameter.
\item quantiles: limits of the credible intervals (5\%, 10\%, 50\%, 90\%, 95\%) for
each time point.
\item model: Model used for the calculations.
\item mus: Mean parameter values used for the simulations.
\item sigma: Variance-covariance matrix used for the simulations.
}
}
\details{
FitIsoGrowth class
}
\section{Methods (by generic)}{
\itemize{
\item \code{print(StochasticGrowth)}: print of the model

\item \code{plot(StochasticGrowth)}: Growth prediction (prediction band) considering
parameter uncertainty.

}}
