% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{summary.FitDynamicGrowth}
\alias{summary.FitDynamicGrowth}
\title{Summary of a FitDynamicGrowth object}
\usage{
\method{summary}{FitDynamicGrowth}(object, ...)
}
\arguments{
\item{object}{Instance of FitDynamicGrowth}

\item{...}{ignored}
}
\description{
Summary of a FitDynamicGrowth object
}
