% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{summary.FitDynamicGrowthMCMC}
\alias{summary.FitDynamicGrowthMCMC}
\title{Summary of a FitDynamicGrowthMCMC object}
\usage{
\method{summary}{FitDynamicGrowthMCMC}(object, ...)
}
\arguments{
\item{object}{Instance of FitDynamicGrowthMCMC}

\item{...}{ignored}
}
\description{
Summary of a FitDynamicGrowthMCMC object
}
