% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{summary.FitIsoGrowth}
\alias{summary.FitIsoGrowth}
\title{Summary of a FitIsoGrowth object}
\usage{
\method{summary}{FitIsoGrowth}(object, ...)
}
\arguments{
\item{object}{Instance of FitIsoGrowth}

\item{...}{ignored}
}
\description{
Summary of a FitIsoGrowth object
}
