% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{summary.FitMultipleDynamicGrowth}
\alias{summary.FitMultipleDynamicGrowth}
\title{Summary of a FitMultipleDynamicGrowth object}
\usage{
\method{summary}{FitMultipleDynamicGrowth}(object, ...)
}
\arguments{
\item{object}{Instance of FitMultipleDynamicGrowth}

\item{...}{ignored}
}
\description{
Summary of a FitMultipleDynamicGrowth object
}
