% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{summary.FitMultipleGrowthMCMC}
\alias{summary.FitMultipleGrowthMCMC}
\title{Summary of a FitMultipleGrowthMCMC object}
\usage{
\method{summary}{FitMultipleGrowthMCMC}(object, ...)
}
\arguments{
\item{object}{instance of FitMultipleGrowthMCMC.}

\item{...}{ignored.}
}
\description{
Summary of a FitMultipleGrowthMCMC object
}
