% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{summary.FitSecondaryGrowth}
\alias{summary.FitSecondaryGrowth}
\title{Summary of a FitSecondaryGrowth object}
\usage{
\method{summary}{FitSecondaryGrowth}(object, ...)

\method{summary}{FitSecondaryGrowth}(object, ...)
}
\arguments{
\item{object}{Instance of FitSecondaryGrowth}

\item{...}{ignored}
}
\description{
Summary of a FitSecondaryGrowth object

Summary of a FitSecondaryGrowth object
}
