% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_counts.R
\name{time_to_size}
\alias{time_to_size}
\title{Time for the population to reach a given size}
\usage{
time_to_size(model, size, logbase = c("natural", "10"))
}
\arguments{
\item{model}{An instance of \link{IsothermalGrowth}, \link{DynamicGrowth}, \link{StochasticGrowth}
or \link{MCMCgrowth}}

\item{size}{Target population size (in log scale)}
}
\value{
If "model" is \link{IsothermalGrowth} or \link{DynamicGrowth}, a number. If "model"
is \link{StochasticGrowth} or \link{MCMCgrowth}, an instance of \link{TimeDistribution}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculates the elapsed time required for the population to reach a given size
(in log scale)
}
\details{
The calculation method differs depending on the type of object passed. If the
model contains a deterministic prediction (\link{IsothermalGrowth} or \link{DynamicGrowth}),
the time to reach the given population size is estimated by linear interpolation and
the function a number.

If, on the other hand, a stochastic prediction is passed (\link{StochasticGrowth},
\link{MCMCgrowth}), the time to reach "size" is estimated for each Monte Carlo iteration.
Then, the function returns an instance of \link{TimeDistribution}, which contains
the distribution of times.
}
