% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDistribution_class.R
\name{TimeDistribution}
\alias{TimeDistribution}
\alias{plot.TimeDistribution}
\title{TimeDistribution class}
\usage{
\method{plot}{TimeDistribution}(x, y = NULL, ..., bin_width = 1)
}
\arguments{
\item{x}{The object of class \code{TimeDistribution} to plot.}

\item{y}{ignored.}

\item{...}{ignored.}

\item{bin_width}{A number that specifies the width of a bin in the histogram, 
see: \code{\link{geom_histogram}}}
}
\description{
The \code{TimeDistribution} class contains an estimate of the probability
distribution of the time to reach a given microbial count. 
Its constructor is \code{\link{distribution_to_logcount}}.

It is a subclass of list with the items:
\itemize{
\item distribution Sample of the distribution of times to reach \code{log_count}.
\item summary Summary statistics of distribution (mean, sd, median, q10 and q90).
}
}
\section{Functions}{
\itemize{
\item \code{plot.TimeDistribution}: plot of the distribution of the time to reach a 
microbial count.
}}

