% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_isothermal_growth.R
\name{trilinear_model}
\alias{trilinear_model}
\title{Trilinear growth model}
\usage{
trilinear_model(times, logN0, mu, lambda, logNmax)
}
\arguments{
\item{times}{Numeric vector of storage times}

\item{logN0}{Initial log microbial count}

\item{mu}{Maximum specific growth rate}

\item{lambda}{Lag phase duration}

\item{logNmax}{Maximum log microbial count}
}
\value{
Numeric vector with the predicted microbial count.
}
\description{
Trilinear growth model defined by Buchanan et al. (1997).
}
