% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{vcov.FitDynamicGrowth}
\alias{vcov.FitDynamicGrowth}
\title{Variance matrix of a dynamic fit}
\usage{
\method{vcov}{FitDynamicGrowth}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitDynamicGrowth}.}

\item{...}{ignored}
}
\description{
Returns the unscaled covariance matrix, calculated as 1/(0.5*Hessian)
}
