% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{vcov.FitDynamicGrowthMCMC}
\alias{vcov.FitDynamicGrowthMCMC}
\title{Variance matrix of an MCMC dynamic fit}
\usage{
\method{vcov}{FitDynamicGrowthMCMC}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitDynamicGrowthMCMC}.}

\item{...}{ignored}
}
\description{
Returns the covariance, estimated as the variance of the samples
from the Markov chain.
}
