% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{vcov.FitIsoGrowth}
\alias{vcov.FitIsoGrowth}
\title{Variance matrix of an isothermal fit}
\usage{
\method{vcov}{FitIsoGrowth}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitIsoGrowth}.}

\item{...}{ignored}
}
\description{
Returns the unscaled covariance matrix, calculated as 1/(0.5*Hessian)
}
