% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{vcov.FitMultipleDynamicGrowth}
\alias{vcov.FitMultipleDynamicGrowth}
\title{Variance matrix of a global fit}
\usage{
\method{vcov}{FitMultipleDynamicGrowth}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitMultipleDynamicGrowth}.}

\item{...}{ignored}
}
\description{
Returns the unscaled covariance matrix, calculated as 1/(0.5*Hessian)
}
