% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{vcov.FitMultipleGrowthMCMC}
\alias{vcov.FitMultipleGrowthMCMC}
\title{Variance matrix of a global fit using MCMC}
\usage{
\method{vcov}{FitMultipleGrowthMCMC}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitMultipleGrowthMCMC}.}

\item{...}{ignored}
}
\description{
Returns the covariance, estimated as the variance of the samples
from the Markov chain.
}
