% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{vcov.FitSecondaryGrowth}
\alias{vcov.FitSecondaryGrowth}
\title{Variance matrix of a cardinal fit}
\usage{
\method{vcov}{FitSecondaryGrowth}(object, ...)
}
\arguments{
\item{object}{an instance of \code{FitSecondaryGrowth}.}

\item{...}{ignored}
}
\description{
Returns the unscaled covariance matrix, calculated as 1/(0.5*Hessian)
}
