% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metainfo_secondary.R
\name{check_secondary_pars}
\alias{check_secondary_pars}
\title{Basic checks of secondary parameters}
\usage{
check_secondary_pars(secondary_models)
}
\arguments{
\item{secondary_models}{A list of secondary models returned by extract_secondary_pars}
}
\description{
Checks that the model name is correct, that the number of model parameters is correct
and that every parameter is defined.
}
