% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_fit.R
\name{fit_growth}
\alias{fit_growth}
\title{Fitting microbial growth}
\usage{
fit_growth(
  fit_data,
  model_keys,
  start,
  known,
  environment = "constant",
  algorithm = "regression",
  approach = "single",
  env_conditions = NULL,
  ...,
  check = TRUE,
  logbase = c("natural", "10"),
  formula = logN ~ time
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
