% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_iso_growth.R
\name{fit_isothermal_growth}
\alias{fit_isothermal_growth}
\title{Fit isothermal growth models}
\usage{
fit_isothermal_growth(fit_data, model_name, starting_point, known_pars, ...)
}
\description{
Fit isothermal growth models
}
