% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_multi_dynamic.R
\name{fit_multiple_growth_MCMC}
\alias{fit_multiple_growth_MCMC}
\title{Fitting growth models to multiple dynamic experiments using MCMC}
\usage{
fit_multiple_growth_MCMC(
  starting_point,
  experiment_data,
  known_pars,
  sec_model_names,
  niter,
  ...
)
}
\arguments{
\item{experiment_data}{a nested list with the experimental data. Each entry describes
one experiment as a list with two elements: data and conditions. \code{data} is a tibble
with two columns: time and logN. \code{conditions} is a tibble with one column named time
and as many additional columns as environmental factors.}

\item{known_pars}{named vector of known model parameters}

\item{sec_model_names}{named character vector with names the
environmental conditions and values the secondary model (see secondary_model_data).}

\item{niter}{number of samples of the MCMC algorithm.}

\item{...}{additional arguments for \code{modMCMC}.}
}
\value{
An instance of
}
\description{
This functions enables to fit a growth model using a dataset comprised of
several experiments with potentially different dynamic experimental conditions.
}
