% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_secondary_models.R
\name{fit_secondary_growth}
\alias{fit_secondary_growth}
\title{Fit secondary growth models}
\usage{
fit_secondary_growth(
  fit_data,
  starting_point,
  known_pars,
  sec_model_names,
  transformation = "sq",
  ...
)
}
\description{
Fit secondary growth models
}
