% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dynamic_growth.R
\name{get_dyna_residuals}
\alias{get_dyna_residuals}
\title{Residuals of dynamic prediction}
\usage{
get_dyna_residuals(
  this_p,
  fit_data,
  env_conditions,
  known_pars,
  sec_model_names,
  cost = NULL,
  logbase_mu = logbase_logN,
  logbase_logN = 10
)
}
\arguments{
\item{this_p}{named vector of model parameters}

\item{fit_data}{tibble with the data for the fit}

\item{env_conditions}{tibble with the environmental conditions}

\item{known_pars}{named vector of known model parameters}

\item{sec_model_names}{named character vector with names the
environmental conditions and values the secondary model (e.g. 'CPM').}

\item{cost}{an instance of modCost to be combined (to fit multiple models).}

\item{logbase_mu}{Base of the logarithm the growth rate is referred to.
By default, 10 (i.e. log10). See vignette about units for details.}
}
\value{
An instance of \code{\link[=modCost]{modCost()}}.
}
\description{
Function for calculating residuals of a dynamic prediction according
to the requirements of \code{\link[=modFit]{modFit()}}.
}
