% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_secondary_models.R
\name{get_secondary_residuals}
\alias{get_secondary_residuals}
\title{Residuals of secondary models}
\usage{
get_secondary_residuals(
  this_p,
  my_data,
  known_pars,
  sec_model_names,
  transformation
)
}
\arguments{
\item{this_p}{Named vector of model parameter values.}

\item{my_data}{Tibble with the data used for the fit.}

\item{known_pars}{Named vector of fixed model paramaters.}

\item{sec_model_names}{Named character vector defining the secondary model
for each environmental factor.}

\item{transformation}{Character defining the tranformation of \code{mu} for
model fitting. One of \code{sq} (square root), \code{log} (log-transform) or
\code{none} (no transformation).}
}
\value{
A numeric vector of residuals.
}
\description{
Residual function for \code{\link{fit_secondary_growth}}.
}
