% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalGrowthComparison_class.R
\name{GlobalGrowthComparison}
\alias{GlobalGrowthComparison}
\alias{coef.GlobalGrowthComparison}
\alias{summary.GlobalGrowthComparison}
\alias{print.GlobalGrowthComparison}
\alias{plot.GlobalGrowthComparison}
\title{GlobalGrowthComparison class}
\usage{
\method{coef}{GlobalGrowthComparison}(object, ...)

\method{summary}{GlobalGrowthComparison}(object, ...)

\method{print}{GlobalGrowthComparison}(x, ...)

\method{plot}{GlobalGrowthComparison}(x, y, ..., type = 1, add_trend = TRUE)
}
\arguments{
\item{object}{an instance of GlobalGrowthComparison}

\item{...}{ignored}

\item{x}{an instance of GlobalGrowthComparison}

\item{y}{ignored}

\item{type}{if type==1, the plot compares the model predictions. If type ==2,
the plot compares the parameter estimates. If type==3, the plot shows the residuals}

\item{add_trend}{should a trend line of the residuals be added for type==3? \code{TRUE}
by default}
}
\description{
The \code{GlobalGrowthComparison} class contains several functions for model comparison
and model selection of growth models. It should not be instanced directly. Instead,
it should be constructed using \code{\link[=compare_growth_fits]{compare_growth_fits()}}. It is similar to
\link{GrowthComparison}, although with specific tools to deal with several experiments.

It includes two type of tools for model selection and comparison: statistical indexes
and visual analyses. Please check the sections below for details.

Note that all these tools use the names defined in \code{\link[=compare_growth_fits]{compare_growth_fits()}}, so
we recommend passing a named list to that function.
}
\section{Methods (by generic)}{
\itemize{
\item \code{coef(GlobalGrowthComparison)}: table of parameter estimates

\item \code{summary(GlobalGrowthComparison)}: summary table for the comparison

\item \code{print(GlobalGrowthComparison)}: print of the model comparison

\item \code{plot(GlobalGrowthComparison)}: illustrations comparing the fitted models

}}
\section{Statistical indexes}{

\code{GlobalGrowthComparison} implements two S3 methods to obtain numerical values to facilitate
model comparison and selection.
\itemize{
\item the \code{coef} method returns a tibble with the values of the parameter estimates
and their corresponding standard errors for each model.
\item the \code{summary} returns a tibble with the AIC, number of degrees of freedom,
mean error and root mean squared error for each model.
}
}

\section{Visual analyses}{

The S3 plot method can generate three types of plots:
\itemize{
\item when \code{type = 1}, the plot compares the fitted growth curves against the
experimental data used to fit the model.
\item when \code{type = 2}, the plot compares the parameter estimates using error bars, where
the limits of the error bars are the expected value +/- one standard error. In case
one model does not has some model parameter (i.e. either because it is not defined
or because it was fixed), the parameter is not included in the plot.
\item when \code{type=3}, the plot shows the tendency of the residuals for each model. This
plot can be used to detect deviations from independence.
}

These plots are divided by facets for each experiment.
}

