% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsothermalGrowth_class.R
\name{IsothermalGrowth}
\alias{IsothermalGrowth}
\alias{print.IsothermalGrowth}
\alias{plot.IsothermalGrowth}
\alias{coef.IsothermalGrowth}
\title{IsothermalGrowth class}
\usage{
\method{print}{IsothermalGrowth}(x, ...)

\method{plot}{IsothermalGrowth}(
  x,
  y = NULL,
  ...,
  line_col = "black",
  line_size = 1,
  line_type = "solid",
  ylims = NULL,
  label_y = NULL,
  label_x = "time"
)

\method{coef}{IsothermalGrowth}(object, ...)
}
\arguments{
\item{x}{The object of class \link{IsothermalGrowth} to plot.}

\item{...}{ignored}

\item{y}{ignored}

\item{line_col}{Aesthetic parameter to change the colour of the line,
see: \code{\link[=geom_line]{geom_line()}}}

\item{line_size}{Aesthetic parameter to change the thickness of the line,
see: \code{\link[=geom_line]{geom_line()}}}

\item{line_type}{Aesthetic parameter to change the type of the line,
takes numbers (1-6) or strings ("solid") see: \code{\link[=geom_line]{geom_line()}}}

\item{ylims}{Two-dimensional numeric vector with the limits of the y-axis
(or \code{NULL}, which is the default)}

\item{label_y}{Title of the y-axis}

\item{label_x}{Title of the x-axis}

\item{object}{an instance of \link{IsothermalGrowth}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

The class \link{IsothermalGrowth} has been superseded by the top-level
class \link{GrowthPrediction}, which provides a unified approach for growth modelling.

Still, it is still returned if the superseded \code{\link[=predict_isothermal_growth]{predict_isothermal_growth()}} is called.

It is a subclass of list with the items:
\itemize{
\item simulation: A tibble with the model simulation.
\item model: The name of the model used for the predictions.
\item pars: A list with the values of the model parameters.
}
}
\section{Methods (by generic)}{
\itemize{
\item \code{print(IsothermalGrowth)}: print of the model

\item \code{plot(IsothermalGrowth)}: plot of the predicted growth curve.

\item \code{coef(IsothermalGrowth)}: coefficients of the model

}}
