% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lambda_to_Q0}
\alias{lambda_to_Q0}
\title{Q0 from lag phase duration}
\usage{
lambda_to_Q0(lambda, mu, logbase_mu = 10)
}
\arguments{
\item{lambda}{Duration of the lag phase.}

\item{mu}{Specific growth rate in the exponential phase.}

\item{logbase_mu}{Base of the logarithm the growth rate is referred to.
By default, 10 (i.e. log10). See vignette about units for details.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Convenience function to calculate the value of Q0 for the Baranyi model from
the duration of the lag phase
}
