% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCgrowth_class.R
\name{MCMCgrowth}
\alias{MCMCgrowth}
\alias{plot.MCMCgrowth}
\title{MCMCgrowth class}
\usage{
\method{plot}{MCMCgrowth}(x, y = NULL, ...)
}
\arguments{
\item{x}{The object of class \code{MCMCgrowth} to plot.}

\item{y}{ignored}

\item{...}{ignored.}
}
\description{
The \code{MCMCgrowth} class contains the results of a growth prediction
consider parameter variability based on a model
fitted using \code{\link{fit_MCMC_growth}} or 
\code{\link{fit_multiple_growth_MCMC}}.

It is a subclass of list with items:
\itemize{
\item sample: Parameter sample used for the calculations.
\item simulations: Individual growth curves calculated based on the parameter
sample.
\item quantiles: Tibble with the limits of the credible intervals
 (5%, 10%, 50%, 90% and 95%) for each time point.
\item model: Instance of \code{FitDynamicGrowthMCMC} used for predictions.
}
}
\section{Functions}{
\itemize{
\item \code{plot.MCMCgrowth}: plot of predicted growth (prediction band).
}}

