% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_plotting.R
\name{plot.DynamicGrowth}
\alias{plot.DynamicGrowth}
\title{Plot of DynamicGrowth Object}
\usage{
\method{plot}{DynamicGrowth}(
  x,
  y = NULL,
  ...,
  add_factor = NULL,
  ylims = NULL,
  label_y1 = "logN",
  label_y2 = add_factor
)
}
\arguments{
\item{x}{The object of class \code{DynamicGrowth} to plot.}

\item{y}{ignored}

\item{...}{additional arguments passed to \code{plot}.}

\item{add_factor}{whether to plot also one environmental factor.
If \code{NULL} (default), no environmenta factor is plotted. If set
to one character string that matches one entry of x$env_conditions,
that condition is plotted in the secondary axis}

\item{ylims}{A two dimensional vector with the limits of the primary y-axis.}

\item{label_y1}{Label of the primary y-axis.}

\item{label_y2}{Label of the secondary y-axis.}
}
\value{
An instance of \code{ggplot}.
}
\description{
Plots the predicted log microbial count for an
instance of \code{DynamicGrowth}.
}
