% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_plotting.R
\name{plot.FitIsoGrowth}
\alias{plot.FitIsoGrowth}
\title{Plot of FitIsoGrowth Object}
\usage{
\method{plot}{FitIsoGrowth}(
  x,
  y = NULL,
  ...,
  line_col = "black",
  line_size = 0.5,
  line_type = 1,
  point_col = "black",
  point_size = 0.5,
  point_shape = 16
)
}
\arguments{
\item{x}{The object of class \code{FitIsoGrowth} to plot.}

\item{y}{ignored}

\item{...}{additional arguments passed to \code{plot}.}

\item{line_col}{Aesthetic parameter to change the colour of the line geom in the plot, see: \code{\link{geom_line}}}

\item{line_size}{Aesthetic parameter to change the thickness of the line geom in the plot, see: \code{\link{geom_line}}}

\item{line_type}{Aesthetic parameter to change the type of the line geom in the plot, takes numbers (1-6) or strings ("solid") see: \code{\link{geom_line}}}

\item{point_col}{Aesthetic parameter to change the colour of the point geom, see: \code{\link{geom_point}}}

\item{point_size}{Aesthetic parameter to change the size of the point geom, see: \code{\link{geom_point}}}

\item{point_shape}{Aesthetic parameter to change the shape of the point geom, see: \code{\link{geom_point}}}
}
\value{
An instance of \code{ggplot}.
}
\description{
Plots the predicted log microbial count for an
instance of \code{FitIsoGrowth}.
}
