% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dynamic_growth.R
\name{predict_dynamic_growth}
\alias{predict_dynamic_growth}
\title{Growth under dynamic conditions}
\usage{
predict_dynamic_growth(times, env_conditions, primary_pars, secondary_models)
}
\arguments{
\item{times}{Numeric vector of storage times to make the predictions}

\item{env_conditions}{Tibble describing the variation of the environmental
conditions during storage. It must have a column named \code{time} with the
storage time and as many additional columns as environmental factors.}

\item{primary_pars}{A named list defining the parameters of the primary model
and the initial values of the model variables. That is, with names \code{mu_opt},
\code{Nmax}, \code{N0}, \code{Q0}.}

\item{secondary_models}{A nested list describing the secondary models.}
}
\value{
A list of class \code{DynamicGrowth} with items:
\itemize{
\item simulation: A tibble with the model prediction
\item gammas: A tibble with the value of each gamma factor for each
value of \code{times}.
\item env_conditions: A list of functions interpolating the environmental
conditions.
\item primary_pars: A list with the model parameters of the primary model.
\item sec_models: A nested list defining the secondary models.
}
}
\description{
Predicts microbial growth under dynamic conditions based on the
Baranyi model (Baranyi and Roberts, 1994) and secondary models
based on the gamma concept (Zwietering et al. 1992).
}
\details{
Model predictions are done by linear interpolation of the environmental
conditions defined in \code{env_conditions}.
}
