% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_MCMC_growth.R
\name{predict_MCMC_growth}
\alias{predict_MCMC_growth}
\title{Stochastic growth of MCMC fit}
\usage{
predict_MCMC_growth(MCMCfit, times, env_conditions, niter)
}
\arguments{
\item{MCMCfit}{An instance of \code{FitDynamicGrowthMCMC}.}

\item{times}{Numeric vector of storage times for the predictions.}

\item{env_conditions}{Tibble with the (dynamic) environmental conditions
during the experiment. It must have one column named 'time' with the
storage time and as many columns as required with the environmental conditions.}

\item{niter}{Number of iterations.}
}
\value{
A list of class \code{MCMCgrowth} with items:
\itemize{
\item sample: Parameter sample used for the calculations.
\item simulations: Individual growth curves calculated based on the parameter
sample.
\item quantiles: Tibble with the limits of the credible intervals
 (5%, 10%, 50%, 90% and 95%) for each time point.
\item model: Instance of \code{FitDynamicGrowthMCMC} used for predictions.
}
}
\description{
Makes a stochastic prediction of microbial growth based on a growth model
fitted using \code{\link{fit_MCMC_growth}}. This function predicts growth
curves for \code{niter} samples (with replacement) of the samples of the
MCMC algorithm. Then, credible intervals are calculated based on the
quantiles of the model predictions at each time point.
}
