% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_stochastic_growth.R
\name{predict_stochastic_growth}
\alias{predict_stochastic_growth}
\title{Isothermal growth with variability}
\usage{
predict_stochastic_growth(
  model_name,
  times,
  n_sims,
  mean_logN0,
  sd_logN0,
  mean_sqmu,
  sd_sqmu,
  mean_sqlambda,
  sd_sqlambda,
  mean_logNmax,
  sd_logNmax,
  corr_matrix = diag(4)
)
}
\description{
Isothermal growth with variability
}
