% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Q0_to_lambda}
\alias{Q0_to_lambda}
\title{Lag phase duration from Q0}
\usage{
Q0_to_lambda(q0, mu, logbase_mu = 10)
}
\arguments{
\item{q0}{Initial value of the variable Q.}

\item{mu}{Specific growth rate in the exponential phase.}

\item{logbase_mu}{Base of the logarithm the growth rate is referred to.
By default, 10 (i.e. log10). See vignette about units for details.}
}
\description{
Convenience function to calculate the lag phase duration (lambda) of the
Baranyi model from the
maximum specific growth rate and the initial value of the variable Q.
}
\details{
Note that this function uses the unit system of biogrowth (i.e. log10). Care must
be taken when using parameters obtained from other sources.
}
