% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R
\name{residuals.FitMultipleGrowthMCMC}
\alias{residuals.FitMultipleGrowthMCMC}
\title{Residuals of FitMultipleGrowthMCMC}
\usage{
\method{residuals}{FitMultipleGrowthMCMC}(object, ...)
}
\arguments{
\item{object}{Instance of FitMultipleGrowthMCMC}

\item{...}{ignored}
}
\description{
Residuals of FitMultipleGrowthMCMC
}
