% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_coupled.R
\name{show_guess_coupled}
\alias{show_guess_coupled}
\title{Plot of the initial guess for the Baranyi-Ratkowsky model}
\usage{
show_guess_coupled(
  fit_data,
  guess,
  mode = "two_steps",
  logbase_mu = exp(1),
  logbase_logN = 10
)
}
\arguments{
\item{fit_data}{Tibble (or data.frame) of data for the fit. It must have two columns, one with
the elapsed time (\code{time} by default) and another one with the decimal logarithm
of the populatoin size (\code{logN} by default). Different column names can be
defined using the \code{formula} argument.}

\item{guess}{Named vector with the initial guess of the model parameters}

\item{logbase_mu}{Base of the logarithm the growth rate is referred to.
By default, 10 (i.e. log10). See vignette about units for details.}

\item{model_name}{Character defining the primary growth model as per \code{\link[=primary_model_data]{primary_model_data()}}}

\item{formula}{an object of class "formula" describing the x and y variables.
\code{logN ~ time} as a default.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} comparing the model prediction against the data
}
\description{
Compares the prediction corresponding to a guess of the parameters of the Baranyi-Ratkowsky
model against experimental data
}
