% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_counts.R
\name{time_to_logcount}
\alias{time_to_logcount}
\title{Time to reach a given microbial count}
\usage{
time_to_logcount(model, log_count)
}
\arguments{
\item{model}{An instance of \code{IsothermalGrowth} or \code{DynamicGrowth}.}

\item{log_count}{The target log microbial count.}
}
\value{
The predicted time to reach \code{log_count}.
}
\description{
Returns the storage time required for the microbial count to
reach \code{log_count} according to the predictions of \code{model}.
Calculations are done using linear interpolation of the model predictions.
}
